requirements.txt
pandas
numpy
python-dotenv
mysql-connector-python
pymysql
cryptography
scikit-learn
#-e .